LayoutTests/platform/gtk/imported/w3c/web-platform-tests/mathml/presentation-markup/scripts/subsup-6-expected.txt


FAIL mprescripts coordinates and sizes (2 postscripts) assert_approx_equals: 2 postscripts, top expected 51 +/- 1 but got 16
FAIL mprescripts coordinates and sizes (2 prescripts) assert_approx_equals: 2 prescripts, left expected 43 +/- 1 but got 8
FAIL mprescripts coordinates and sizes (2 prescripts, 4 postscripts) assert_approx_equals: 2 prescripts, 4 postscripts, left expected 43 +/- 1 but got 8
FAIL mprescripts coordinates and sizes (4 prescripts, 2 postscripts) assert_approx_equals: 4 prescripts, 2 postscripts, left expected 78 +/- 1 but got 8
FAIL mprescripts coordinates and sizes (prescripts with padding/border/margin) assert_approx_equals: prescripts with padding/border/margin, left expected 43 +/- 1 but got 8
